[ALPS07882006] camsys: fix coverity issues
1. unsigend compared with 0: no effect
2. array index violation
3. fix reserved_i2c time

MTK-Commit-Id: 15f1a2fd0e251eb8d2a289e7bd2d7cccdd36ba22

CR-Id: ALPS07882006
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7s/cam/mtk_cam-job_state.c
@@ -77,7 +77,7 @@ const char *str_state(int state_type, int state)
 	if (WARN_ON((unsigned int)state_type >= ARRAY_SIZE(str_fn)))
 		return "(not-found)";
 
-	return str_fn[state_type](state);
+	return str_fn[(unsigned int)state_type](state);
 }
 
 const char *str_state_type(int state_type)
